+++ Sources/Sentry/SentrySDK.m
@@ -58,6 +58,8 @@ NS_ASSUME_NONNULL_BEGIN
 @implementation SentrySDK
 static SentryHub *_Nullable currentHub;
 static NSObject *currentHubLock;
+static SentryLogger *_Nullable currentLogger;
+static NSObject *currentLoggerLock;
 static BOOL crashedLastRunCalled;
 static SentryAppStartMeasurement *sentrySDKappStartMeasurement;
 static NSObject *sentrySDKappStartMeasurementLock;
@@ -81,6 +83,7 @@ static NSDate *_Nullable startTimestamp = nil;
     if (self == [SentrySDK class]) {
         sentrySDKappStartMeasurementLock = [[NSObject alloc] init];
         currentHubLock = [[NSObject alloc] init];
+        currentLoggerLock = [[NSObject alloc] init];
         startOptionsLock = [[NSObject alloc] init];
         startInvocations = 0;
         _detectedStartUpCrash = NO;
@@ -112,6 +115,26 @@ static NSDate *_Nullable startTimestamp = nil;
     return replay;
 }
 #endif
+
++ (SentryLogger *)logger
+{
+    @synchronized(currentLoggerLock) {
+        if (currentLogger == nil) {
+            SentryLogBatcher *batcher;
+            if (nil != currentHub.client && currentHub.client.options.experimental.enableLogs) {
+                batcher = [[SentryLogBatcher alloc] initWithClient:currentHub.client];
+            } else {
+                batcher = nil;
+            }
+            currentLogger = [[SentryLogger alloc]
+                 initWithHub:currentHub
+                dateProvider:SentryDependencyContainer.sharedInstance.dateProvider
+                     batcher:batcher];
+        }
+        return currentLogger;
+    }
+}
+
 /** Internal, only needed for testing. */
 + (void)setCurrentHub:(nullable SentryHub *)hub
 {
@@ -629,6 +652,10 @@ static NSDate *_Nullable startTimestamp = nil;
     [hub bindClient:nil];
 
     [SentrySDK setCurrentHub:nil];
+
+    @synchronized(currentLoggerLock) {
+        currentLogger = nil;
+    }
 
     [SentryCrashWrapper.sharedInstance stopBinaryImageCache];
     [SentryDependencyContainer.sharedInstance.binaryImageCache stop];