+++ Sources/Sentry/SentryWatchdogTerminationTrackingIntegration.m
@@ -79,9 +79,15 @@ NS_ASSUME_NONNULL_BEGIN
 
     [self.tracker start];
 
+#    if SDK_V9
+    BOOL isV2Enabled = YES;
+#    else
+    BOOL isV2Enabled = options.enableAppHangTrackingV2;
+#    endif // SDK_V9
+
     self.anrTracker =
         [SentryDependencyContainer.sharedInstance getANRTracker:options.appHangTimeoutInterval
-                                                    isV2Enabled:options.enableAppHangTrackingV2];
+                                                    isV2Enabled:isV2Enabled];
     [self.anrTracker addListener:self];
 
     self.appStateManager = appStateManager;