+++ CHANGELOG.md
@@ -2,9 +2,15 @@
 
 ## Unreleased
 
+### Features
+
+- Add experimental support for capturing structured logs via `SentrySDK.logger` (#5532)
+
 ### Improvements
 
 - Extract video processing to a new class (#5604)
+- Move continuous profiling payload serialization off of the main thread (#5613)
+- Improve video generation using apple recommended loop (#5612)
 
 ### Fixes